Refactored data and terminology to its own component
